LPD-66489 Rename hooks to useDLVideoExternalShortcutFields
