+++ modules/apps/document-library/document-library-video/src/main/resources/META-INF/resources/js/utils/useDLVideoExternalShortcutFields.ts
